[WEJBHTTP-142] Refactoring - Renaming EjbConstants to Constants.
