+++ ejb/src/main/java/org/wildfly/httpclient/ejb/HttpCancelHandler.java
@@ -18,7 +18,7 @@
 
 package org.wildfly.httpclient.ejb;
 
-import static org.wildfly.httpclient.ejb.EjbConstants.JSESSIONID_COOKIE_NAME;
+import static org.wildfly.httpclient.ejb.Constants.JSESSIONID_COOKIE_NAME;
 
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
